refactor out nav bar
